requirements.txt
asgiref==3.7.2
Django==4.2.5
types-pytz==2023.3.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
urllib3==2.0.4
PyYAML==6.0.1
